changed urls, error version

+++ src/main/java/com/pw/cinema/movie_category/MovieCategoryService.java
@@ -1,6 +1,7 @@
 package com.pw.cinema.movie_category;
 
 import com.pw.cinema.exceptions.AlreadyExistsException;
+import com.pw.cinema.exceptions.HasMoviesException;
 import com.pw.cinema.movie.Movie;
 import com.pw.cinema.movie.MovieRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,17 +14,17 @@ import java.util.Map;
 @Service
 public class MovieCategoryService {
 
-    @Autowired
-    private MovieCategoryRepository movieCategoryRepository;
+    private final MovieCategoryRepository movieCategoryRepository;
 
-    @Autowired
-    private MovieRepository movieRepository;
+    public MovieCategoryService(MovieCategoryRepository movieCategoryRepository) {
+        this.movieCategoryRepository = movieCategoryRepository;
+    }
 
-    public Object getCategory(MovieCategory movieCategory) {
-        MovieCategory movieCategory1 = movieCategoryRepository.findById(movieCategory.getId()).get();
+    public Object getCategory(Long id) {
+        MovieCategory movieCategory1 = movieCategoryRepository.findById(id).get();
         Map<String, Object> response = new HashMap<>();
         response.put("data", movieCategory1);
-        response.put("message", "Successfully found movies");
+        response.put("message", "Successfully found category");
         response.put("success", true);
         return response;
     }
@@ -32,7 +33,7 @@ public class MovieCategoryService {
         List<MovieCategory> movieCategoryList = movieCategoryRepository.findAll();
         Map<String, Object> response = new HashMap<>();
         response.put("data", movieCategoryList);
-        response.put("message", "Successfully found movies");
+        response.put("message", "Successfully found all categories");
         response.put("success", true);
         return response;
     }
@@ -44,36 +45,35 @@ public class MovieCategoryService {
         }
         Map<String, Object> response = new HashMap<>();
         response.put("data", movieCategoryRepository.save(movieCategory));
-        response.put("message", "Successfully added new movie");
+        response.put("message", "Successfully added new category");
         response.put("success", true);
         return response;
     }
 
 
     public Object updateCategory(MovieCategory categoryChanges, Long id) {
+        MovieCategory movieCategory = movieCategoryRepository.findById(id).get();
+        movieCategory.setName(categoryChanges.getName());
+        movieCategory.setSlug(categoryChanges.getSlug());
+        // movies add or change
         Map<String, Object> response = new HashMap<>();
-        response.put("data", movieCategoryRepository.save(categoryChanges));
-        response.put("message", "Successfully found movies");
+        response.put("data", movieCategoryRepository.save(movieCategory));
+        response.put("message", "Successfully updated category");
         response.put("success", true);
         return response;
     }
 
-    public Object deleteCategory(Long id) {
+    public Object deleteCategory(Long id) throws HasMoviesException {
         Map<String, Object> response = new HashMap<>();
+        MovieCategory movieCategory = movieCategoryRepository.findById(id).get();
+        boolean isUsed = movieCategory.getMovies().size() > 0;
+        if (isUsed) {
+            throw new HasMoviesException("This category can not be deleted, because it includes movies");
+        }
         movieCategoryRepository.deleteById(id);
-        response.put("message", "Successfully deleted movie");
+        response.put("message", "Successfully deleted this category");
         response.put("success", true);
         return response;
-    }
-
-    public Object getAllMoviesByCategoryId(Long categoryId) {
-        List <Movie> movies = movieRepository.findMoviesByMovieCategoriesId(categoryId);
-        return movies;
-    }
-
-    public Object getAllMovieCategoriesByMovieId(Long movieId) {
-        List <MovieCategory> movieCategories = movieCategoryRepository.findMovieCategoriesByMoviesId(movieId);
-        return movieCategories;
     }
 
 }
